dapp: make raiden service class more independent
The RaidenService class is closely tight to the Web3Provider and
ConfigProvider classes. At best it should not even know those exist.
Rather a higher level ("dirty") module should orchestrate the usage
between them. This is important as in future there will be different
provider options with according configurations.
To loose the components, the RaidenService now takes all necessary
values - including the ethereum provider instance - as parameters. It is
now the Home view components job to get the provider and configuration
to connect the SDK. Later it is possible to opt in as user to specify
different providers and configurations with visual components.