dapp: unify ethereum provider to json rpc provider
The custom EthereumProvider type was a tool to act between the type
conjunction of the SDKs interface and the actual connection methods as
used by the dApp. This could be simplified to make sure that all
connection methods return an JsonRpcProvider. In fact the last not
unified method was the String as direct URL of an Json RPC server.
Within the SDK this is a one-liner to pack it into a JsonRpcProvider
object. As it (heavily) simplifies the typing in the dApp it is good to
just do the packing right here. This is especially attractive when
having the planned connection manager in mind.